Run strongSwan as non-root user
Add capabilities to charon and iptables to be able to run strongSwan as
a non-root user.

Adding CAP_NET_ADMIN to iptables is required for strongSwan to still be
able to create firewall rules as needed.

A new user 'strongswan' is created at build time which will later run
the daemon.
It is highly recommended to create a new user for strongSwan on the host
to map into the container at runtime. (e.g. docker run --user=
[ user | user:group | uid | uid:gid | user:gid | uid:group ])

Doing so helps to protect secrets such as private keys by changing their
ownership to the host strongSwan user.